pyproject.toml
[tool.poetry]
name = "avaliacao-jota"
version = "0.1.0"
description = ""
authors = ["Paulo S. Garcia <pgarciapydev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.6"
djangorestframework = "^3.15.2"
django-cors-headers = "^4.7.0"
django-environ = "^0.12.0"
drf-yasg = "^1.21.8"
djangorestframework-simplejwt = "^5.4.0"
celery = "^5.4.0"
redis = "^5.2.1"
psycopg2-binary = "^2.9.10"
gunicorn = "^23.0.0"
pillow = "^11.1.0"


[tool.poetry.group.dev.dependencies]
django-extensions = "^3.2.3"
black = "^25.1.0"
isort = "^6.0.0"
flake8 = "^7.1.2"
pytest = "^8.3.4"
pytest-django = "^4.10.0"


[tool.poetry.group.test.dependencies]
pytest = "^8.3.4"
pytest-django = "^4.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
